requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
contextlib2==21.6.0
Flask==2.1.1
future==0.18.2
hvac==0.10.11
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.24.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
threadloop==1.0.2
thrift==0.16.0
tornado==5.1.1
tzdata==2023.3
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
